- decoupled content Script (opt-out-ext.js)
