+++ src/content/functions/onError.js
@@ -1,0 +1,7 @@
+/**
+ * @description Function handles errors.
+ * @param error
+ */
+export default function (error) {
+  console.error(`Error: ${error}`);
+}
