+++ src/content/functions/processTweets.js
@@ -1,0 +1,11 @@
+import checkText from './checkText';
+// TODO: Check opt-out-ext.js for changes needed here
+export default function (selector, popupPrefs) {
+  const posts = document.querySelectorAll(selector); // selecting tweet object
+  posts.forEach((post) => {
+    if (post.classList.contains('processed-true')) return;
+    if (post.classList.contains('processed-false')) return;
+    if (post.classList.contains('processing')) return;
+    checkText(post, selector, popupPrefs);
+  });
+}
